+++ platform/services/services.environment/src/main/java/de/iip_ecosphere/platform/services/environment/services/TransportToAasConverter.java
@@ -66,6 +66,8 @@ public abstract class TransportToAasConverter<T> {
     private long timeout = 20 * 60 * 1000; // cleanup after 20 minutes
     private long lastCleanup = System.currentTimeMillis();
     private long cleanupTimeout = 5 * 1000; // when the next cleanup shall be considered
+    private long aasFailedTimestamp = -1;
+    private long aasFailedTimeout = 30 * 1000;
     private TraceRecordReceptionCallback callback;
     private List<IOConsumer<T>> notifier = new ArrayList<>();
     private Supplier<Boolean> aasEnabledSupplier = () -> true;
@@ -159,6 +161,15 @@ public abstract class TransportToAasConverter<T> {
      */
     public void setTimeout(long timeout) {
         this.timeout = timeout;
+    }
+
+    /**
+     * Changes the cleanup timeout, i.e., the time between two cleanups.
+     * 
+     * @param cleanupTimeout the timeout in ms
+     */
+    public void setCleanupTimeout(long cleanupTimeout) {
+        this.cleanupTimeout = cleanupTimeout;
     }
     
     /**
@@ -257,7 +268,11 @@ public abstract class TransportToAasConverter<T> {
      */
     protected void handleNew(T data) {
         // add new record
-        if (isAasEnabled()) {
+        long now = System.currentTimeMillis();
+        if (isAasEnabled() && aasFailedTimestamp > 0 && now - aasFailedTimestamp < aasFailedTimeout) {
+            aasFailedTimestamp = -1; // allow for re-try
+        }
+        if (isAasEnabled() && aasFailedTimestamp < 0) {
             try {
                 if (null == aas) {
                     // do not populate, we just add/remove in this class
@@ -273,6 +288,7 @@ public abstract class TransportToAasConverter<T> {
                 cleanup(aas);
             } catch (IOException e) {
                 LoggerFactory.getLogger(getClass()).error("Cannot obtain AAS {}: {}", getAasUrn(), e.getMessage());
+                aasFailedTimestamp = now;
             }
         }
     }
@@ -400,14 +416,30 @@ public abstract class TransportToAasConverter<T> {
     }
     
     /**
+     * Pursues a cleanup of the (internally known) AAS.
+     * 
+     * @return whether a cleanup process was executed (not whether elements were deleted)
+     */
+    public boolean cleanup() {
+        boolean done = false;
+        if (null != aas) {
+            done = cleanup(aas);
+        }
+        return done;
+    }
+    
+    /**
      * Cleans up outdated trace entries. Called in {@link #handleNew(Object)} if regular input is expected,
      * may be called regularly by an external timer.
      * 
      * @param aas the AAS to clean up
+     * @return whether a cleanup process was executed (not whether elements were deleted)
      * @see #getCleanupPredicate()
+     * @see #cleanup()
      */
-    public void cleanup(Aas aas) {
+    public boolean cleanup(Aas aas) {
         // remove outdated ones
+        boolean done = false;
         long now = System.currentTimeMillis();
         if (now - lastCleanup > cleanupTimeout) {
             long timestamp = now - timeout;
@@ -424,7 +456,9 @@ public abstract class TransportToAasConverter<T> {
                 return cont;
             }, SubmodelElementCollection.class);
             lastCleanup = now;
+            done = true;
         }
+        return done;
     }
 
     /**