+++ platform/documentation/RELEASES.md
@@ -22,12 +22,9 @@ In progress:
 * [#70](https://github.com/iip-ecosphere/platform/issues/70): UI progress reporting 
 * [#117](https://github.com/iip-ecosphere/platform/issues/117): Multiple service managers per device
 * [#146](https://github.com/iip-ecosphere/platform/issues/146): UI editors for configuration tables
-* [#145](https://github.com/iip-ecosphere/platform/issues/145): UI table for nameplate configurations
 * [#144](https://github.com/iip-ecosphere/platform/issues/144): Deployment plans and application instances
 * [#143](https://github.com/iip-ecosphere/platform/issues/143): Container testing mode(s)
-* [#142](https://github.com/iip-ecosphere/platform/issues/142): AAS resilience for ECS-Runtime/Service Manager
 * [#141](https://github.com/iip-ecosphere/platform/issues/141): Fix AppAAS performance/resource consumption issue
-* [#62](https://github.com/iip-ecosphere/platform/issues/62): Platform component heartbeat and AAS cleanup
 * [#115](https://github.com/iip-ecosphere/platform/issues/115): Cleanup, e.g., old build processes
 
 New features:
@@ -35,8 +32,11 @@ New features:
 * [#100](https://github.com/iip-ecosphere/platform/issues/100): Allowing to start/stop multiple app instances (only via deployment plans)
 * [#59](https://github.com/iip-ecosphere/platform/issues/59): Basic version of distributed testing/evaluation environment for platform and apps (PETE)
 * [#90](https://github.com/iip-ecosphere/platform/issues/90): UI display application service meshes
+* [#62](https://github.com/iip-ecosphere/platform/issues/62): Platform component heartbeat and AAS cleanup
+* [#142](https://github.com/iip-ecosphere/platform/issues/142): AAS resilience for ECS-Runtime/Service Manager
 * [#107](https://github.com/iip-ecosphere/platform/issues/107): Platform installation scripts
 * [#56](https://github.com/iip-ecosphere/platform/issues/56): Network-based integration protocol for Python service environment (websockets)
+* [#145](https://github.com/iip-ecosphere/platform/issues/145): UI table for nameplate configurations
 * Generic time series data aggregator in data functions
 * Application intercommunication support (`AppIntercom`)
 * [#132](https://github.com/iip-ecosphere/platform/issues/132): Bitmotec system monitoring integration
@@ -54,6 +54,7 @@ Improvements:
 * Connector mocking data with $period and $repeats, preparation for exact JSON field mapping
 * Transport connector can now have a local serialization provider, e.g., for MQTT out.
 * Trace-to-AAS service and (new) extensible Java service can utilize optional (output) transport connector, e.g., for northbound integration.
+* More reliable artifact file modification watching, in particular for Linux.
 * Build process fixes for examples.
 
 ### Release 0.5.0 (2023/03/01):